+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Use temp directory with Star-Fusion
 - Resource bump for RNA
+- Limit memory for glnexus
 
 ## [10.2.3]